refactor redis_int_tests.
